+++ addon/utils/layer-to-jsts.js
@@ -1,7 +1,9 @@
 import jsts from 'npm:jsts';
+import Ember from 'ember';
 import { latLngToCoords, latLngsToCoords } from './lat-lng-to-coord';
 
 let geometryFactory = new jsts.geom.GeometryFactory();
+let geometryReducer = null;
 
 // convert coordinates in jsts object
 let coordinatesFunction = function(coord, altitude) {
@@ -11,9 +13,12 @@ let coordinatesFunction = function(coord, altitude) {
 };
 
 // for point
-let latlngToPointJsts = function(latlng, crs, precision) {
+let latlngToPointJsts = function(latlng, crs, precision, scale) {
   let coord = latLngToCoords(latlng, crs, precision, coordinatesFunction);
   let point = geometryFactory.createPoint(coord);
+  if (!Ember.isNone(scale)) {
+    point = geometryPrecisionReducer(scale).reduce(point);
+  }
 
   if (point) {
     point.setSRID(crs.code.split(':')[1]);
@@ -27,7 +32,7 @@ let coordinatesToLineString = function (coordinates) {
   return geometryFactory.createLineString(coordinates);
 };
 
-let latlngToPolylineJsts = function (latlngs, crs, precision) {
+let latlngToPolylineJsts = function (latlngs, crs, precision, scale) {
   let multi = !L.LineUtil.isFlat(latlngs);
   let coords = latLngsToCoords(latlngs, crs, multi ? 1 : 0, false, precision, coordinatesFunction);
 
@@ -37,6 +42,10 @@ let latlngToPolylineJsts = function (latlngs, crs, precision) {
   } else {
     let geometries = coords.map(coordinatesToLineString);
     polyline = geometryFactory.createMultiLineString(geometries);
+  }
+
+  if (!Ember.isNone(scale)) {
+    polyline = geometryPrecisionReducer(scale).reduce(polyline);
   }
 
   if (polyline) {
@@ -60,7 +69,7 @@ let coordinatesToPolygon = function(coordinates) {
   }
 };
 
-let latlngToPolygonJsts = function(latlngs, crs, precision) {
+let latlngToPolygonJsts = function(latlngs, crs, precision, scale) {
   let holes = !L.LineUtil.isFlat(latlngs);
   let multi = holes && !L.LineUtil.isFlat(latlngs[0]);
   let coords = latLngsToCoords(latlngs, crs, multi ? 2 : holes ? 1 : 0, true, precision, coordinatesFunction);
@@ -75,6 +84,10 @@ let latlngToPolygonJsts = function(latlngs, crs, precision) {
   } else {
     let geometries = coords.map(coordinatesToPolygon);
     polygon = geometryFactory.createMultiPolygon(geometries);
+  }
+
+  if (!Ember.isNone(scale)) {
+    polygon = geometryPrecisionReducer(scale).reduce(polygon);
   }
 
   if (polygon) {
@@ -95,19 +108,25 @@ let coordToJsts = function(coord) {
 };
 
 let geometryPrecisionReducer = function(scale) {
-  return new jsts.precision.GeometryPrecisionReducer(new jsts.geom.PrecisionModel(scale));
+  if (geometryReducer === null) {
+    geometryReducer = new jsts.precision.GeometryPrecisionReducer(new jsts.geom.PrecisionModel(scale));
+  }
+
+  return geometryReducer;
 };
 
-let geometryToJsts = function(geometry) {
+let geometryToJsts = function(geometry, scale) {
   let coords = [];
   switch (geometry.type) {
     case 'Point':
       coords = coordToJsts(geometry.coordinates);
-      return geometryFactory.createPoint(coords);
+      let point = geometryFactory.createPoint(coords);
+      return Ember.isNone(scale) ? point : geometryPrecisionReducer(scale).reduce(point);
     case 'LineString':
       let coordsAsArray = geometry.coordinates instanceof Array ? geometry.coordinates : Array.from(geometry.coordinates);
       coords = coordsAsArray.map(coordToJsts);
-      return coordinatesToLineString(coords);
+      let line = coordinatesToLineString(coords);
+      return Ember.isNone(scale) ? line : geometryPrecisionReducer(scale).reduce(line);
     case 'MultiLineString':
       for (let i = 0; i < geometry.coordinates.length; i++) {
         let coordsAsArray = geometry.coordinates[i] instanceof Array ? geometry.coordinates[i] : Array.from(geometry.coordinates[i]);
@@ -115,14 +134,16 @@ let geometryToJsts = function(geometry) {
       }
 
       let geometries = coords.map(coordinatesToLineString);
-      return geometryFactory.createMultiLineString(geometries);
+      let multiLine = geometryFactory.createMultiLineString(geometries);
+      return Ember.isNone(scale) ? multiLine : geometryPrecisionReducer(scale).reduce(multiLine);
     case 'Polygon':
       for (let i = 0; i < geometry.coordinates.length; i++) {
         let coordsAsArray = geometry.coordinates[i] instanceof Array ? geometry.coordinates[i] : Array.from(geometry.coordinates[i]);
         coords.push(coordsAsArray.map(coordToJsts));
       }
 
-      return coordinatesToPolygon(coords);
+      let polygon = coordinatesToPolygon(coords);
+      return Ember.isNone(scale) ? polygon : geometryPrecisionReducer(scale).reduce(polygon);
     case 'MultiPolygon':
       for (let i = 0; i < geometry.coordinates.length; i++) {
         let coordinates = [];
@@ -135,7 +156,8 @@ let geometryToJsts = function(geometry) {
       }
 
       let geom = coords.map(coordinatesToPolygon);
-      return geometryFactory.createMultiPolygon(geom);
+      let multiPolygon = geometryFactory.createMultiPolygon(geom);
+      return Ember.isNone(scale) ? multiPolygon : geometryPrecisionReducer(scale).reduce(multiPolygon);
   }
 };