+++ CHANGELOG.md
@@ -11,7 +11,8 @@ All notable changes to this project will be documented in this file.
 * Add support for point objects (markers) in vector layers legends.
 * Add pane and renderer on objects in layer WFS when layer is created.
 * Add utils open\close sub menu in componetns: feature-result-item and flexberry-maplayers.
-* Add geometryPrecisionReducer in function getMulti
+* Add geometryPrecisionReducer in functions utils/layer-to-jsts
+* Add constant scale on mapApi
 
 ### Changed
 * Change chart's library from `highcharts` to `chart.js`.